+++ decoder/LAVVideo/parsers/MPEG2HeaderParser.h
@@ -24,7 +24,7 @@ struct GetBitContext;
 class CMPEG2HeaderParser
 {
 public:
-  CMPEG2HeaderParser(const BYTE *pData, int length);
+  CMPEG2HeaderParser(const BYTE *pData, size_t length);
   ~CMPEG2HeaderParser(void);
 
 public:
@@ -39,7 +39,7 @@ public:
   } hdr;
 
 private:
-  void ParseMPEG2Header(const BYTE *pData, int length);
+  void ParseMPEG2Header(const BYTE *pData, size_t length);
   void MPEG2ParseSequenceHeader(GetBitContext *gb);
   void MPEG2ParseExtHeader(GetBitContext *gb);
 };
