+++ demuxer/LAVSplitter/LAVSplitter.cpp
@@ -340,13 +340,13 @@ STDMETHODIMP CLAVSplitter::GetSite(REFIID riid, void **ppvSite)
 STDMETHODIMP_(int) CLAVSplitter::GetCount()
 {
   CAutoLock pinLock(&m_csPins);
-  return m_pPins.size();
+  return (int)m_pPins.size();
 }
 
 STDMETHODIMP CLAVSplitter::GetStatus(int i, int& samples, int& size)
 {
   CAutoLock pinLock(&m_csPins);
-  if (i >= m_pPins.size())
+  if ((size_t)i >= m_pPins.size())
     return E_FAIL;
 
   CLAVOutputPin *pPin = m_pPins.at(i);
